merge Grokability master

+++ app/Http/Controllers/Users/UsersController.php
@@ -56,7 +56,7 @@ class UsersController extends Controller
     public function create(Request $request)
     {
         $this->authorize('create', User::class);
-        $groups = Group::pluck('name', 'id');
+        $groups = Group::orderBy('name', 'asc')->pluck('name', 'id');
 
         $userGroups = collect();
 
@@ -122,18 +122,31 @@ class UsersController extends Controller
         // Strip out the superuser permission if the user isn't a superadmin
         $permissions_array = $request->input('permission');
 
-        if (! auth()->user()->isSuperUser()) {
-            unset($permissions_array['superuser']);
+        // Strip out the individual superuser permission if the API user isn't a superadmin
+        if (!auth()->user()->isSuperUser()) {
+
+            if ((is_array($permissions_array)) && (array_key_exists('superuser', $permissions_array))) {
+                unset($permissions_array['superuser']);
+            }
         }
+
+        // Strip out the individual admin permission if the API user isn't an admin
+        if (!auth()->user()->isAdmin()) {
+
+            if ((is_array($permissions_array)) && (array_key_exists('admin', $permissions_array))) {
+                unset($permissions_array['admin']);
+            }
+        }
+
         $user->permissions = json_encode($permissions_array);
 
         // we have to invoke the form request here to handle image uploads
         app(ImageUploadRequest::class)->handleImages($user, 600, 'avatar', 'avatars', 'avatar');
 
-        if ($request->get('redirect_option') === 'back'){
+        if ($request->input('redirect_option') === 'back'){
             session()->put(['redirect_option' => 'index']);
         } else {
-            session()->put(['redirect_option' => $request->get('redirect_option')]);
+            session()->put(['redirect_option' => $request->input('redirect_option')]);
         }
 
 
@@ -151,7 +164,9 @@ class UsersController extends Controller
             }
 
             if ($request->filled('groups')) {
-                $user->groups()->sync($request->input('groups'));
+                if (auth()->user()->can('canEditAuthFields', $user) && auth()->user()->can('editableOnDemo')) {
+                    $user->groups()->sync($request->input('groups'));
+                }
             } else {
                 $user->groups()->sync([]);
             }
@@ -199,7 +214,7 @@ class UsersController extends Controller
             }
 
             $permissions = config('permissions');
-            $groups = Group::pluck('name', 'id');
+            $groups = Group::orderBy('name', 'asc')->pluck('name', 'id');
 
             $userGroups = $user->groups()->pluck('name', 'id');
             $user->permissions = $user->decodePermissions();
@@ -325,7 +340,7 @@ class UsersController extends Controller
 
         // Handle uploaded avatar
         app(ImageUploadRequest::class)->handleImages($user, 600, 'avatar', 'avatars', 'avatar');
-        session()->put(['redirect_option' => $request->get('redirect_option')]);
+        session()->put(['redirect_option' => $request->input('redirect_option')]);
 
         if ($user->save()) {
             // Redirect to the user page
@@ -351,10 +366,13 @@ class UsersController extends Controller
         if ($user = User::find($id)) {
 
             $this->authorize('delete', $user);
+            if (auth()->user()->can('canEditAuthFields', $user) && auth()->user()->can('editableOnDemo')) {
 
-            if ($user->delete()) {
-                return redirect()->route('users.index')->with('success', trans('admin/users/message.success.delete'));
+                if ($user->delete()) {
+                    return redirect()->route('users.index')->with('success', trans('admin/users/message.success.delete'));
+                }
             }
+            return redirect()->route('users.index')->with('error', trans('admin/users/message.cannot_delete'));
         }
         return redirect()->route('users.index')->with('error', trans('admin/users/message.user_not_found'));
 
@@ -505,7 +523,8 @@ class UsersController extends Controller
     public function getExportUserCsv()
     {
         $this->authorize('view', User::class);
-        \Debugbar::disable();
+
+        $this->disableDebugbar();
 
         $response = new StreamedResponse(function () {
             // Open output stream
